requirements.txt
apache-airflow==2.9.3  # Certifique-se de que a versão seja compatível com seu setup
pandas==2.1.4          # Instala a versão mais recente do pandas
